assets: featured communities styling
* reorganise the file structure due to local development symlink issues
