--- site/cds_rdm/assets/semantic-ui/less/cds_rdm/.gitkeep
@@ -1,1 +1,0 @@
-# This file exists only to add less/ folder to git repository.
