--- site/cds_rdm/assets/semantic-ui/less/my_site/.gitkeep
@@ -1,1 +1,0 @@
-# This file exists only to add less/ folder to git repository.
