+++ site/setup.cfg
@@ -10,6 +10,7 @@ tests =
 [options.entry_points]
 flask.commands =
     cds-admin = cds_rdm.cli:cds_admin
+    migration = cds_rdm.migration.cli:migration
 invenio_base.blueprints =
     cds_rdm_views = cds_rdm.views:create_blueprint
 invenio_assets.webpack =
@@ -18,5 +19,3 @@ cds_rdm.migrator.models =
     cms_note = cds_rdm.migration.transform.models.note:model
 cds_rdm.migrator.rules =
     base = cds_rdm.migration.transform.xml_processing.rules
-flask.commands =
-    migration = cds_rdm.migration.cli:migration
